@@ -1,6 +1,5 @@
 package mate.academy.spring.service.mapper;
 
-import java.util.stream.Collectors;
 import mate.academy.spring.dto.response.ShoppingCartResponseDto;
 import mate.academy.spring.model.ShoppingCart;
 import mate.academy.spring.model.Ticket;
@@ -17,7 +16,7 @@ public class ShoppingCartMapper implements
         responseDto.setTicketIds(shoppingCart.getTickets()
                 .stream()
                 .map(Ticket::getId)
-                .collect(Collectors.toList()));
+                .toList());
         return responseDto;
     }
 }
